refacto: éclate les services front Transformateurs par étapes
