+++ .eslintrc.js
@@ -35,4 +35,16 @@ module.exports = {
       flowVersion: "0.53", // Flow version
     },
   },
+  overrides: [
+    {
+      // or whatever matches stories specified in .storybook/main.js
+      files: ["*.stories.@(ts|tsx|js|jsx|mjs|cjs)"],
+      rules: {
+        // example of overriding a rule
+        "storybook/hierarchy-separator": "error",
+        // example of disabling a rule
+        "storybook/default-exports": "off",
+      },
+    },
+  ],
 };
