+++ anssi-nis2-ui/package.json
@@ -60,7 +60,6 @@
     "eslint-plugin-react": "^7.33.2",
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-react-refresh": "^0.4.3",
-    "eslint-plugin-storybook": "^0.6.15",
     "jsdom": "^22.1.0",
     "prettier": "3.0.3",
     "storybook": "^7.5.0",
