+++ anssi-nis2-ui/src/Components/Simulateur/Etapes/EtapeActivites.tsx
@@ -1,16 +1,14 @@
 import { FormSimulateur } from "./index.ts";
 import React, { useMemo } from "react";
 import { AssociationSectorielleActivite } from "../../../Domaine/Simulateur/ActivitesParSecteurEtSousSecteur.ts";
-import {
-  collecteTitresPourActivite,
-  fabriqueCartographieEntreesLegendeEtOptionsChampSimlulateur,
-} from "../../../Services/Simulateur/Transformateurs.ts";
 import { libellesSecteursActivite } from "../../../References/LibellesSecteursActivite.ts";
 import { libellesSousSecteursActivite } from "../../../References/LibellesSousSecteursActivite.ts";
 
 import { EnsembleChamps } from "../Inputs/EnsembleChamps.tsx";
 
 import { SimulateurContenuEtapeProps } from "../../../Services/Simulateur/Props/simulateurEtapeProps";
+import { fabriqueCartographieEntreesLegendeEtOptionsChampSimlulateur } from "../../../Services/Simulateur/Transformateurs/FabriqueConstructeurOptionActivite.ts";
+import { collecteTitresPourActivite } from "../../../Domaine/Simulateur/CollecteTitresPourActivite.ts";
 
 const EtapeActivitesCalculee = ({
   propageActionSimulateur,
