+++ anssi-nis2-ui/src/Components/Simulateur/Etapes/EtapeOSE.tsx
@@ -1,11 +1,11 @@
 import { libellesDesigneOSE } from "../../../References/Libelles.ts";
 import { FormSimulateur } from "./index.ts";
 import React, { useMemo } from "react";
-import { transformeReponsesDesigneOSEPourSelect } from "../../../Services/Simulateur/Transformateurs.ts";
 import RadioButtons from "@codegouvfr/react-dsfr/RadioButtons";
 import { fabriqueGestionChangementSimple } from "../../../Services/Simulateur/gestionnaires.ts";
 
 import { SimulateurContenuEtapeProps } from "../../../Services/Simulateur/Props/simulateurEtapeProps";
+import { transformeReponsesDesigneOSEPourSelect } from "../../../Services/Simulateur/Transformateurs/TransformeReponsesDesigneOSEPourSelect.ts";
 
 const EtapeOSECalculee = ({
   donneesFormulaire,
