+++ anssi-nis2-ui/src/Components/Simulateur/Etapes/EtapeSecteursActivite.tsx
@@ -1,6 +1,5 @@
 import Checkbox from "@codegouvfr/react-dsfr/Checkbox";
 import { FormSimulateur } from "./index.ts";
-import { transformeSecteursActiviteVersOptions } from "../../../Services/Simulateur/Transformateurs.ts";
 
 import { libellesSecteursActivite } from "../../../References/LibellesSecteursActivite.ts";
 import { useCallback, useMemo } from "react";
@@ -9,6 +8,7 @@ import React from "react";
 import { SimulateurContenuEtapeProps } from "../../../Services/Simulateur/Props/simulateurEtapeProps";
 import { OptionsChampSimulateur } from "../../../Services/Simulateur/Props/optionChampSimulateur";
 import { SecteurActivite } from "../../../Domaine/Simulateur/SecteursActivite";
+import { transformeSecteursActiviteVersOptions } from "../../../Services/Simulateur/Transformateurs/TransformeSecteursActiviteVersOptions.ts";
 
 const CheckboxWrapper = ({
   legend,
