+++ anssi-nis2-ui/src/Components/Simulateur/Etapes/EtapeTypeStructure.tsx
@@ -2,11 +2,11 @@ import { libellesTypesStructure } from "../../../References/Libelles.ts";
 import RadioButtons from "@codegouvfr/react-dsfr/RadioButtons";
 import { FormSimulateur } from "./index.ts";
 import React, { useMemo } from "react";
-import { transformeTypeStructureVersOptions } from "../../../Services/Simulateur/Transformateurs.ts";
 import { SimulateurEtapeNodeComponent } from "../../../Services/Simulateur/Props/component";
 import { fabriqueGestionChangementSimple } from "../../../Services/Simulateur/gestionnaires.ts";
 
 import { SimulateurContenuEtapeProps } from "../../../Services/Simulateur/Props/simulateurEtapeProps";
+import { transformeTypeStructureVersOptions } from "../../../Services/Simulateur/Transformateurs/TransformeTypeStructureVersOptions.ts";
 
 const EtapeTypeStructureCalculee: SimulateurEtapeNodeComponent = ({
   donneesFormulaire,
