+++ anssi-nis2-ui/src/Services/Simulateur/Reducteurs.ts
@@ -10,11 +10,11 @@ import {
   SecteursAvecSousSecteurs,
   SousSecteurActivite,
 } from "../../Domaine/Simulateur/SousSecteurs";
-import { transformateurSousSecteurActivite } from "./Transformateurs.ts";
 import { entreesLibellesSousSecteurs } from "../../References/LibellesSousSecteursActivite.ts";
 import { OptionsChampSimulateur } from "./Props/optionChampSimulateur";
 import { BoutonsNavigation } from "./Props/boutonsNavigation";
 import { fabriqueSecteurContientLeSousSecteur } from "../../Domaine/Simulateur/Operations/operationsSecteurs.ts";
+import { transformateurSousSecteurActivite } from "./Transformateurs/TransformateurSousSecteurActivite.ts";
 
 const generateNewStateFrom = (
   state: DonneesFormulaireSimulateur,
