+++ anssi-nis2-ui/src/Services/Simulateur/Transformateurs/TransformePaysUnionEuropeennePourSelect.ts
@@ -1,0 +1,13 @@
+import { AppartenancePaysUnionEuropeenne } from "../../../Domaine/Simulateur/ChampsSimulateur";
+import { TransformeRecordToSelect } from "../Operations/optionChampSimulateur";
+import { genereTransformateurValeursVersOptions } from "../genereTransformateurValeursVersOptions.ts";
+
+const getPaysUnionEuropeenneElement = (
+  value: string,
+  paysUnionEuropeenne: Partial<Record<AppartenancePaysUnionEuropeenne, string>>,
+) => paysUnionEuropeenne[value as AppartenancePaysUnionEuropeenne] || value;
+export const transformePaysUnionEuropeennePourSelect: TransformeRecordToSelect<AppartenancePaysUnionEuropeenne> =
+  genereTransformateurValeursVersOptions(
+    getPaysUnionEuropeenneElement,
+    "etatMembre",
+  );
