+++ anssi-nis2-ui/src/Services/Simulateur/Transformateurs/TransformeSecteursActiviteVersOptions.ts
@@ -1,0 +1,13 @@
+import { SecteurActivite } from "../../../Domaine/Simulateur/SecteursActivite";
+import { TransformeRecordToSelect } from "../Operations/optionChampSimulateur";
+import { genereTransformateurValeursVersOptions } from "../genereTransformateurValeursVersOptions.ts";
+
+export const getSecteurActiviteLabel = (
+  value: string,
+  secteurActivite: Partial<Record<SecteurActivite, string>>,
+) => secteurActivite[value as SecteurActivite] || value;
+export const transformeSecteursActiviteVersOptions: TransformeRecordToSelect<SecteurActivite> =
+  genereTransformateurValeursVersOptions(
+    getSecteurActiviteLabel,
+    "secteurActivite",
+  );
