+++ anssi-nis2-ui/src/Services/Simulateur/Transformateurs/TransformeTypeStructureVersOptions.ts
@@ -1,0 +1,13 @@
+import { TypeStructure } from "../../../Domaine/Simulateur/ChampsSimulateur";
+import { TransformeRecordToSelect } from "../Operations/optionChampSimulateur";
+import { genereTransformateurValeursVersOptions } from "../genereTransformateurValeursVersOptions.ts";
+
+const getTypesStructureElement = (
+  value: string,
+  typesStructure: Partial<Record<TypeStructure, string>>,
+) => typesStructure[value as TypeStructure] || value;
+export const transformeTypeStructureVersOptions: TransformeRecordToSelect<TypeStructure> =
+  genereTransformateurValeursVersOptions(
+    getTypesStructureElement,
+    "typeStructure",
+  );
