+++ anssi-nis2-ui/test/Domaine/Activites.spec.ts
@@ -2,11 +2,9 @@ import { describe, expect, it } from "vitest";
 import { DonneesFormulaireSimulateur } from "../../src/Domaine/Simulateur/DonneesFormulaire";
 import { libellesSecteursActivite } from "../../src/References/LibellesSecteursActivite";
 import { libellesSousSecteursActivite } from "../../src/References/LibellesSousSecteursActivite";
-import {
-  cartographieSousSecteursParSecteur,
-  collecteTitresPourActivite,
-} from "../../src/Services/Simulateur/Transformateurs";
 import { AssociationSectorielleActivite } from "../../src/Domaine/Simulateur/ActivitesParSecteurEtSousSecteur";
+import { collecteTitresPourActivite } from "../../src/Domaine/Simulateur/CollecteTitresPourActivite";
+import { cartographieSousSecteursParSecteur } from "../../src/Domaine/Simulateur/CartographieSousSecteursParSecteur";
 
 describe("Questionnaire activités", () => {
   it("Construit un tableau avec les sous-secteurs sélectionnés remplaçant le secteur correspondant", () => {
